updating entities Sold and Store

+++ src/main/java/com/unitri/comp/ecommerce/model/entity/Sold.java
@@ -2,31 +2,31 @@ package com.unitri.comp.ecommerce.model.entity;
 
 public class Sold {
 
-    private Long id;
-    private Long product_id;
+    private int id;
+    private int product_id;
     private int quantity;
     private double price;
 
-    public Sold(long id, long product_id, int quantity, double price) {
+    public Sold(int id, int product_id, int quantity, double price) {
         this.id = id;
         this.product_id = product_id;
         this.quantity = quantity;
         this.price = price;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
-    public Long getProduct_id() {
+    public int getProduct_id() {
         return product_id;
     }
 
-    public void setProduct_id(Long product_id) {
+    public void setProduct_id(int product_id) {
         this.product_id = product_id;
     }
